Create nxpprog pip package
Resolves #29
