+++ __main__.py
@@ -1,0 +1,9 @@
+import signal
+import sys
+
+from . import nxpprog
+
+
+if __name__ == '__main__':
+    signal.signal(signal.SIGINT, )
+    sys.exit(nxpprog.main())
